Refactor: Move location of StackFrame
